move suggestion utils into their own file
