+++ platforms/web/lib/suggestion.ts
@@ -1,0 +1,46 @@
+/*
+Copyright 2023 The Matrix.org Foundation C.I.C.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+import { SuggestionPattern } from '../generated/wysiwyg';
+import { SUGGESTIONS } from './constants';
+import { MappedSuggestion, SuggestionChar, SuggestionType } from './types';
+
+function getSuggestionChar(suggestion: SuggestionPattern): SuggestionChar {
+    return SUGGESTIONS[suggestion.key];
+}
+
+function getSuggestionType(suggestion: SuggestionPattern): SuggestionType {
+    switch (suggestion.key) {
+        case 0:
+        case 1:
+            return 'mention';
+        case 2:
+            return 'command';
+        default:
+            return 'unknown';
+    }
+}
+
+export function mapSuggestion(
+    suggestion: SuggestionPattern | null,
+): MappedSuggestion | null {
+    if (suggestion === null) return suggestion;
+    return {
+        ...suggestion,
+        keyChar: getSuggestionChar(suggestion),
+        type: getSuggestionType(suggestion),
+    };
+}
